fix: wrap session search params in suspense
